requirements.txt
aio-pika
aiormq
alembic
annotated-types
anyio
asyncpg
bcrypt
black
certifi
cryptography
fast-depends
fastapi
faststream
getenv
greenlet
httpcore
httpx
idna
iniconfig
Mako
MarkupSafe
multidict
mypy-extensions
packaging
pamqp
pathspec
platformdirs
pluggy
propcache
pycodestyle
pycparser
pydantic
pydantic-settings
pydantic_core
PyJWT
pytest
pytest-asyncio
python-dotenv
python-multipart
sniffio
SQLAlchemy
starlette
typing_extensions
uvicorn
